Added target options to lanes using get_version_number
It appears the get_version_number helper method now requires the target
to be specified otherwise it's asking it through the CLI. Of course, this
is not working on Bitrise.

To fix the issue, it's now possible to pass the target as an environment
variable and option (the options having priority if defined).